Minor fix in FuzzyString

+++ Packages/com.dreamlordsdigital.fuzzystring/Runtime/ApproximatelyEquals.cs
@@ -22,7 +22,7 @@ namespace FuzzyString
 			{
 				if (source.Length == target.Length)
 				{
-					comparisonResults.Add(source.HammingDistance(target)/target.Length);
+					comparisonResults.Add(source.HammingDistance(target)/(double)target.Length);
 				}
 			}
